backend-aionix/requirements.txt
fastapi
uvicorn
python-dotenv
# Change: Use native Google GenAI SDK instead of OpenAI SDK
google-genai 
httpx
python-multipart 
pydantic
requests